pyproject.toml
[project]
name = "telegram-link-summarizer-agent"
version = "0.1.0"
description = "An agentic Telegram bot to summarize links and papers."
readme = "README.md"
requires-python = ">=3.11" # Adjusted to match Docker image
dependencies = [
    "baml-py>=0.88.0",
    "langgraph>=0.0.57",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.7", # Assuming OpenAI, change if needed
    "python-telegram-bot[ext]>=21.0",
    "requests>=2.31.0",
    "pypdf>=4.2.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.3.3", # Adding Tavily as it's in config
    # Add any other specific langchain community/experimental packages if used
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langgraph-cli[inmem]>=0.2.7",
    "marimo>=0.13.2",
    "langchain-community>=0.3.22",
    "rich>=14.0.0",
    "loguru>=0.7.3",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "pymupdf>=1.25.5",
    "click>=8.1.8",
    "h11>=0.16.0",
    "starlette>=0.46.2",
    "pydantic>=2.11.3",
    "pydantic-core>=2.33.1",
    "typing-extensions>=4.13.2",
    "typing-inspection>=0.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.9.0",
    "nest-asyncio>=1.6.0",
    "yt-dlp>=2025.4.30",
    "youtube-transcript-api>=1.1.0",
    "google-api-python-client>=2.169.0",
    "google-cloud-secret-manager>=2.20.0", # Added for GCP Secret Manager access
    "bs4>=0.0.2",
    "playwright>=1.52.0",
    "agentql>=1.10.0",
    "beautifulsoup4>=4.13.4",
    "supabase>=2.16.0",
    "google-genai>=1.73.1",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
    "openai>=1.76.0",
]

[tool.setuptools]
py-modules = ["agent", "bot", "config", "db", "commands", "url_normalize", "summarizer", "personal", "draft_mode", "reminders", "exporter", "export_formatter"]
packages = ["baml_client", "tools", "src", "src.providers", "src.ai", "src.ai.prompts"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=7.1.0",
    "psycopg[binary]>=3.1",
    "respx>=0.23.1",
]
